move modular/namespaced to same directory to fix TS
