+++ packages/functions/package.json
@@ -5,7 +5,7 @@
   "description": "React Native Firebase - Cloud Functions for Firebase lets you automatically run backend code in response to events triggered by Firebase features and HTTPS requests. React Native Firebase supports integration with production and locally emulated Cloud Functions with a simple API interface.\n\n",
   "main": "./dist/commonjs/index.js",
   "module": "./dist/module/index.js",
-  "types": "./dist/typescript/commonjs/functions/lib/modular/index.d.ts",
+  "types": "./dist/typescript/commonjs/functions/lib/index.d.ts",
   "codegenConfig": {
     "name": "NativeFunctionsModule",
     "type": "modules",
@@ -52,13 +52,13 @@
   },
   "exports": {
     ".": {
-      "source": "./lib/modular/index.ts",
+      "source": "./lib/index.ts",
       "import": {
-        "types": "./dist/typescript/module/functions/lib/modular/index.d.ts",
+        "types": "./dist/typescript/module/functions/lib/index.d.ts",
         "default": "./dist/module/index.js"
       },
       "require": {
-        "types": "./dist/typescript/commonjs/functions/lib/modular/index.d.ts",
+        "types": "./dist/typescript/commonjs/functions/lib/index.d.ts",
         "default": "./dist/commonjs/index.js"
       }
     },
@@ -72,7 +72,7 @@
     "!**/__mocks__"
   ],
   "react-native-builder-bob": {
-    "source": "lib/modular",
+    "source": "lib",
     "output": "dist",
     "targets": [
       [
